+++ apps/website/lib/feature/news/ui/news_component.dart
@@ -3,8 +3,8 @@ import 'dart:async';
 import 'package:collection/collection.dart';
 import 'package:common_data/news.dart';
 import 'package:conference_2024_website/gen/i18n/strings.g.dart';
-import 'package:conference_2024_website/ui/home/notifier/news_data_notifier.dart';
-import 'package:conference_2024_website/ui/home/notifier/news_data_state.dart';
+import 'package:conference_2024_website/feature/news/data/news_data_notifier.dart';
+import 'package:conference_2024_website/feature/news/data/news_data_state.dart';
 import 'package:conference_2024_website/ui/theme/extension/theme_extension.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
